Fixed reboot logic, fixed run script exit logic.

+++ run.sh
@@ -25,17 +25,20 @@ sudo /etc/init.d/vmware-workstation-server start
 # Build the boxes.
 packer build -var "box_version=$VERSION" -parallel=false magma-centos7.json
 if [[ $? != 0 ]]; then
-  exit 0
+  exit 1
 else 
   sleep 60
 fi
 packer build -var "box_version=$VERSION" -parallel=false magma-centos6.json
 if [[ $? != 0 ]]; then
-  exit 0
+  exit 1
 else
   sleep 60
 fi
 packer build -var "box_version=$VERSION" -parallel=false magma.json
+if [[ $? != 0 ]]; then
+  exit 1
+fi
 
 # Cleanup the artifacts.
 rm -rf xpti.dat compreg.dat VBoxSVC.log VirtualBox.xml VirtualBox.xml-prev packer_cache/
